+++ packages/werkbank/lib/src/_internal/src/localizations/src/strings_en.g.dart
@@ -254,8 +254,9 @@ class TranslationsAddonsDescriptionEn {
 	final Translations _root; // ignore: unused_field
 
 	// Translations
-	String get component => 'Component Description';
-	String get folder => 'Folder Description';
+	String component({required Object name}) => '${name} (Component)';
+	String folder({required Object name}) => '${name} (Folder)';
+	String get sections => '(Sections)';
 	String get tags => 'Tags';
 	String get description => 'Description';
 	String get links => 'External Links';
@@ -901,8 +902,9 @@ extension on Translations {
 			case 'addons.colorPicker.controls.colorPicker.values.noSelectedColor': return 'No color selected';
 			case 'addons.colorPicker.controls.colorPicker.values.colorCopied': return 'Hex Code copied to clipboard';
 			case 'addons.colorPicker.controls.colorPicker.values.pickedColor': return 'Picked Color';
-			case 'addons.description.component': return 'Component Description';
-			case 'addons.description.folder': return 'Folder Description';
+			case 'addons.description.component': return ({required Object name}) => '${name} (Component)';
+			case 'addons.description.folder': return ({required Object name}) => '${name} (Folder)';
+			case 'addons.description.sections': return '(Sections)';
 			case 'addons.description.tags': return 'Tags';
 			case 'addons.description.description': return 'Description';
 			case 'addons.description.links': return 'External Links';