+++ packages/werkbank/lib/src/addons/src/description/src/metadata/description.dart
@@ -3,26 +3,31 @@ import 'package:werkbank/src/werkbank_internal.dart';
 
 class Descriptions extends UseCaseMetadataEntry<Descriptions> {
   const Descriptions({
+    required this.sectionsDescription,
     required this.folderDescriptions,
     required this.componentDescription,
     required this.useCaseDescription,
   });
 
   static const empty = Descriptions(
-    folderDescriptions: IList.empty(),
+    sectionsDescription: null,
+    folderDescriptions: [],
     componentDescription: null,
     useCaseDescription: null,
   );
 
-  final IList<String> folderDescriptions;
-  final String? componentDescription;
-  final String? useCaseDescription;
+  final DescriptionEntry<WerkbankSections>? sectionsDescription;
+  final List<DescriptionEntry<WerkbankFolder>> folderDescriptions;
+  final DescriptionEntry<WerkbankComponent>? componentDescription;
+  final DescriptionEntry<WerkbankUseCase>? useCaseDescription;
 
   Descriptions copyWith({
-    IList<String>? folderDescriptions,
-    String? componentDescription,
-    String? useCaseDescription,
+    DescriptionEntry<WerkbankSections>? sectionsDescription,
+    List<DescriptionEntry<WerkbankFolder>>? folderDescriptions,
+    DescriptionEntry<WerkbankComponent>? componentDescription,
+    DescriptionEntry<WerkbankUseCase>? useCaseDescription,
   }) => Descriptions(
+    sectionsDescription: sectionsDescription ?? this.sectionsDescription,
     folderDescriptions: folderDescriptions ?? this.folderDescriptions,
     componentDescription: componentDescription ?? this.componentDescription,
     useCaseDescription: useCaseDescription ?? this.useCaseDescription,
@@ -37,24 +42,85 @@ extension DescriptionComposerExtension on UseCaseComposer {
   void description(String description) {
     final currentDescriptions =
         getMetadata<Descriptions>() ?? Descriptions.empty;
+    final node = this.node;
+
+    DescriptionEntry<T> tryMergeIntoEntry<T extends WerkbankNode>({
+      required DescriptionEntry<T>? entry,
+      required T node,
+      required String description,
+    }) {
+      assert(
+        entry == null || entry.node == node,
+        'The node of the entry must match the current node.',
+      );
+      if (entry == null) {
+        return DescriptionEntry<T>(
+          description: description,
+          node: node,
+        );
+      }
+      return DescriptionEntry<T>(
+        description: '${entry.description.trimRight()}\n\n$description',
+        node: node,
+      );
+    }
+
     switch (node) {
       // We treat description on the root as a folder description.
       case WerkbankSections():
-      case WerkbankFolder():
         setMetadata(
           currentDescriptions.copyWith(
-            folderDescriptions: currentDescriptions.folderDescriptions.add(
-              description,
+            sectionsDescription: tryMergeIntoEntry(
+              entry: currentDescriptions.sectionsDescription,
+              node: node,
+              description: description,
             ),
+          ),
+        );
+      case WerkbankFolder():
+        final lastFolder = currentDescriptions.folderDescriptions.lastOrNull;
+        final current = lastFolder?.node == node ? lastFolder : null;
+        final newEntry = tryMergeIntoEntry(
+          entry: current,
+          node: node,
+          description: description,
+        );
+        final newFolderDescriptions = currentDescriptions.folderDescriptions
+            .toList();
+        if (current != null) {
+          newFolderDescriptions[newFolderDescriptions.length - 1] = newEntry;
+        } else {
+          newFolderDescriptions.add(
+            DescriptionEntry<WerkbankFolder>(
+              description: description,
+              node: node,
+            ),
+          );
+        }
+        setMetadata(
+          currentDescriptions.copyWith(
+            folderDescriptions: newFolderDescriptions,
           ),
         );
       case WerkbankComponent():
         setMetadata(
-          currentDescriptions.copyWith(componentDescription: description),
+          currentDescriptions.copyWith(
+            componentDescription: tryMergeIntoEntry(
+              entry: currentDescriptions.componentDescription,
+              node: node,
+              description: description,
+            ),
+          ),
         );
       case WerkbankUseCase():
         setMetadata(
-          currentDescriptions.copyWith(useCaseDescription: description),
+          currentDescriptions.copyWith(
+            useCaseDescription: tryMergeIntoEntry(
+              entry: currentDescriptions.useCaseDescription,
+              node: node,
+              description: description,
+            ),
+          ),
         );
     }
 
@@ -72,4 +138,14 @@ extension DescriptionComposerExtension on UseCaseComposer {
       ),
     );
   }
+}
+
+class DescriptionEntry<T extends WerkbankNode> {
+  DescriptionEntry({
+    required this.description,
+    required this.node,
+  });
+
+  final String description;
+  final T node;
 }