+++ packages/werkbank/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.15.1
+- Make some improvements to use case descriptions
+  - Rename the "About" section for descriptions to "Description".
+  - Calling `c.description(...)` multiple times in the same node will now concatenate the descriptions instead of overwriting them.
+  - The name of a `WerkbankFolder` or `WerkbankComponent` is now displayed above its description.
+
 # 0.15.0
 
 - Initial release